Moving users manager to a new package for user management
