+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/TestCapacitySchedulerNodeLabelUpdate.java
@@ -48,6 +48,8 @@ import org.apache.hadoop.yarn.server.resourcemanager.rmapp.attempt.RMAppAttempt;
 import org.apache.hadoop.yarn.server.resourcemanager.rmcontainer.RMContainerEventType;
 import org.apache.hadoop.yarn.server.resourcemanager.rmcontainer.RMContainerState;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.ResourceScheduler;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.usermanagement.UserInfo;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.usermanagement.UsersManager;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.common.fica.FiCaSchedulerApp;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.event.NodeLabelsUpdateSchedulerEvent;
 import org.apache.hadoop.yarn.server.resourcemanager.webapp.dao.ResourceInfo;
@@ -62,7 +64,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
   private final int GB = 1024;
 
   private YarnConfiguration conf;
-  
+
   RMNodeLabelsManager mgr;
 
   @Before
@@ -73,11 +75,11 @@ public class TestCapacitySchedulerNodeLabelUpdate {
     mgr = new NullRMNodeLabelsManager();
     mgr.init(conf);
   }
-  
+
   private Configuration getConfigurationWithQueueLabels(Configuration config) {
     CapacitySchedulerConfiguration conf =
         new CapacitySchedulerConfiguration(config);
-    
+
     // Define top-level queues
     conf.setQueues(CapacitySchedulerConfiguration.ROOT, new String[] {"a"});
     conf.setCapacityByLabel(CapacitySchedulerConfiguration.ROOT, "x", 100);
@@ -90,7 +92,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
     conf.setCapacityByLabel(A, "x", 100);
     conf.setCapacityByLabel(A, "y", 100);
     conf.setCapacityByLabel(A, "z", 100);
-    
+
     return conf;
   }
 
@@ -144,7 +146,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
     Set<String> set = Sets.newHashSet(elements);
     return set;
   }
-  
+
   private void checkUsedResource(MockRM rm, String queueName, int memory) {
     checkUsedResource(rm, queueName, memory, RMNodeLabelsManager.NO_LABEL);
   }
@@ -269,7 +271,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
       throws Exception {
     // set node -> label
     mgr.addToCluserNodeLabelsWithDefaultExclusivity(ImmutableSet.of("x", "y", "z"));
-    
+
     // set mapping:
     // h1 -> x
     // h2 -> y
@@ -312,17 +314,17 @@ public class TestCapacitySchedulerNodeLabelUpdate {
     containerId2 = ContainerId.newContainerId(am1.getApplicationAttemptId(), 2);
     Assert.assertTrue(rm.waitForState(nm1, containerId2,
         RMContainerState.ALLOCATED));
-    
+
     // check used resource:
     // queue-a used x=1G, ""=1G
     checkUsedResource(rm, "a", 1024, "x");
     checkUsedResource(rm, "a", 1024);
     checkUsedCapacity(rm, "a", 1024, 8000, "x");
     checkUsedCapacity(rm, "a", 1024, 8000);
-    
+
     CapacityScheduler cs = (CapacityScheduler) rm.getResourceScheduler();
     FiCaSchedulerApp app = cs.getApplicationAttempt(am1.getApplicationAttemptId());
-    
+
     // change h1's label to z
     mgr.replaceLabelsOnNode(ImmutableMap.of(nm1.getNodeId(), toSet("z")));
     cs.handle(new NodeLabelsUpdateSchedulerEvent(ImmutableMap.of(nm1.getNodeId(),
@@ -343,7 +345,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
         app.getAppAttemptResourceUsage().getUsed("x").getMemorySize());
     Assert.assertEquals(1024,
         app.getAppAttemptResourceUsage().getUsed("z").getMemorySize());
-    
+
     // change h1's label to y
     mgr.replaceLabelsOnNode(ImmutableMap.of(nm1.getNodeId(), toSet("y")));
     cs.handle(new NodeLabelsUpdateSchedulerEvent(ImmutableMap.of(nm1.getNodeId(),
@@ -370,7 +372,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
         app.getAppAttemptResourceUsage().getUsed("y").getMemorySize());
     Assert.assertEquals(0,
         app.getAppAttemptResourceUsage().getUsed("z").getMemorySize());
-    
+
     // change h1's label to no label
     Set<String> emptyLabels = new HashSet<>();
     Map<NodeId,Set<String>> map = ImmutableMap.of(nm1.getNodeId(),
@@ -412,7 +414,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
         ContainerStatus.newInstance(containerId1, ContainerState.COMPLETE, "",
             ContainerExitStatus.KILLED_BY_RESOURCEMANAGER),
         RMContainerEventType.KILL);
-    
+
     checkUsedResource(rm, "a", 0, "x");
     checkUsedResource(rm, "a", 0, "y");
     checkUsedResource(rm, "a", 0, "z");
@@ -530,7 +532,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
      */
     // set node -> label
     mgr.addToCluserNodeLabelsWithDefaultExclusivity(ImmutableSet.of("x", "y", "z"));
-    
+
     // set mapping:
     // h1 -> x
     // h2 -> y
@@ -572,7 +574,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
         ContainerId.newContainerId(am1.getApplicationAttemptId(), 3);
     Assert.assertTrue(rm.waitForState(nm2, containerId,
         RMContainerState.ALLOCATED));
-    
+
     // app2
     MockRMAppSubmissionData data =
         MockRMAppSubmissionData.Builder.createWithMemory(GB, rm)
@@ -591,7 +593,7 @@ public class TestCapacitySchedulerNodeLabelUpdate {
         ContainerId.newContainerId(am2.getApplicationAttemptId(), 3);
     Assert.assertTrue(rm.waitForState(nm1, containerId,
         RMContainerState.ALLOCATED));
-    
+
     // check used resource:
     // queue-a used x=1G, ""=1G
     checkUsedResource(rm, "a", 3 * GB, "x");
